Rework settings and add time zone
