--- openc3-cosmos-init/plugins/packages/openc3-cosmos-tool-dataextractor/src/tools/DataExtractor/Filters/timeFilters.js
@@ -1,34 +1,0 @@
-/*
-# Copyright 2022 Ball Aerospace & Technologies Corp.
-# All Rights Reserved.
-#
-# This program is free software; you can modify and/or redistribute it
-# under the terms of the GNU Affero General Public License
-# as published by the Free Software Foundation; version 3 with
-# attribution addendums as found in the LICENSE.txt
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Affero General Public License for more details.
-
-# Modified by OpenC3, Inc.
-# All changes Copyright 2022, OpenC3, Inc.
-# All Rights Reserved
-#
-# This file may also be used under the terms of a commercial license 
-# if purchased from OpenC3, Inc.
-*/
-
-export default {
-  filters: {
-    dateTime: function (val, utcOrLocal) {
-      if (utcOrLocal === 'utc') {
-        return val.toISOString()
-      } else {
-        return val.toLocaleString() // TODO: support other locales besides en-US
-      }
-    },
-  },
-  methods: {},
-}
